refactor(): Rename metronome package to metronome_core
Update name to reflect the back-end nature of the files.
